Fixed an instant killing of new-born beech koopa!
a VERY sudden opening:
- in VB6, for loop looks next:
```vb
For(I = 1 To y)
```
which means, the "y" is capturing and constanting inside the loop. In C++, the condition is re-checking same value again physically.
So, to have the same logic as VB, we MUST to declare the copy of a max index:
```cpp
int yMax = y;
for(int I = 1; I <= yMax; I++)
```
So, this solves this damned issue.