Allow players to quit the credits screen by the gamepad
Before this commit the keyboard is required to quit the credits screen. Without of this feature it's the hard challenge for devices with unplugged keyboard such as mobile phones and tablets. When the gamepad is plugged, it's much easier to press the Start button rather taking an effort to get the phone (that was placed far forward at the another table) and touch the BACK button.