Experimental: the World Map performance has been improved
Since the whole age there was full scan of all arrays was been used to find every world map element on the map. This wasn't sensitive for the powerful computers, but it's majorly visible on various weak devices. Just now I implemented the use of the Quad-Tree to resolve this problem and speed-up the whole work.