Fixed the black screen problem on the Android
By the fact, the true reason why the black screen got appeared - the messed up against the texture/screen target left at the incorrect state when entering/exiting the background state. So, make sure the screen target is always set until to call the rendering of the in-game world only.