macOS: Non-debug .app with no assets
Allow the building of the .app with the case it will expect assets at the bundle, but will don't pack them during the build process. This allows creation of the sort of "template" application that can be used to pack multiple games with the same built application. Also, being used to produce the universal build.