gfx_background.cpp: Added the workaround for BG42
There are custom sprites that has the 3455 height which is invalid because it's not multiple 4. That means, if don't have this hack, the sprite will glitch and will be aligned up for one pixel for every animation cycle. Original game had to align this pixel down which fixes thix issue.