Adding Luna Autocode scripting support (#255)
* Cheats: Added call to externally check the buffer

* Initial [NOT BUILDING] work on LunaAutocode

Until start developing the memory emulator, I want to implement the backend of the Luna Autocode engine to get an ability run existng scripts for verifications and debugs.

* The almost working thing

* Ported Calleoca level code

To test out the thing in action

* Fill the NPC memory map

* autocode.cpp: Small tune

* MemEmu: Fixed inability to read-write float fields

* Ouch

* autocode.cpp: Slightly tune the code

* Fixed warnings

* Luna: More fixes and tunes

Also, ported the set of hardcoded level codes for the Analog Funk

* mememu: Tuned the workaround for Science battle

Fixed workflow at Kood's stage where Twomps do fall up or down depending on switch's state

* Luna: Fixed numNPC iterations logic

* MemEMU: Move NPC specific workaround into NPC

TODO: Try to implement an ability to override values

* Added settings path option

(To allow put various custom settings files)

* Added deaths counter and ability to on/off the luna

* Added luna-specific custom cheats

They will work in condition luna engine is enabled

* Modernize the autocode engine

- Get rid of the new/delete, use std::move() when need to move
- Added copy constructor into Autocode class
- Other clean-ups

* Make demos counter work per-game-save

To don't conuse all content of deaths counter, make it save all stuff per game save only

* Update changelog.txt

* compat.cpp: Fixed missing group closing

* Allow deaths counter enabled by compat.ini

For the case episode creator wants to enable the death counter at episode side, add the compat.ini option

* Added path case resolover for Autocode scripts loader

* Luna: Tuned out the render calls

* Small tune of renderOp classes

* Android: Added an ability to run level test by Intent

Using any file manager, you can open one of level files to play

* Remove empty and useless level code

* LunaRender: Small tweaks

* Luna: Added license header and describe level codes

* Move custom level codes engine into own file

* Update Docopoper-TheFloorisLava.cpp

Minor tweak at the comment

* Null checks and small tunes at level codes

* Added second isolated randomizer

For custom stuff

* Polishing the code, also fixing some bugs

- Fixed inability to kill player by command
- Fixed NPC damage printing

* Another small tweak

* And more tunes

* Fixed NPC FaceAll command

Regaring to the original code, the Special6 value is set, not TurnAround

* MemEMU: Added type missmatch warnings

Notify when requested type by mem address is not matching with actual one

* MemEMU: Endianess tweaks

For any cases when it's need to hack exact byte at floating point types, ensure the same format as X86 on the side of the script

* Tune and clean-up the code

* Small polish of level codes

* Updated submodules

* MemEMU: Attempt to fix GCC4.9 error

"implicitly deleted because it's exception-specification does not match the implicit exception-specification"

* MemEMU: Second attempt to fix the GCC 4.9 build

* gfx.cpp: Another GCC 4.9 fix

* Update gfx.cpp

Ouch

* Added an ability to toggle luna engine by compat.ini

* compat.cpp": const enum maps

* Don't run built-in level codes by default

Until the compat.ini enables them

* Update SDL-Mixer-X

* Move all luna-related files into script/luna/

* MemEMU: Added missing byte type check

* Added new event cancel command

It allows to break a set of chained events

* Track all recently called events

* Autocode: Add OnEvent and CancelSMBXEvent

OnEvent - Allows calling an Autocode event
CancelSMBXEvent - allows stop the looped chain of events

* layers.cpp: Fixed an event skipping after removal

* Autocode: optimize the commands enumeration

* autocode_manager.cpp: Small tweak

* More polishing of autocode engine

- Optimize the search of specific command
- Proper script error reporting
- Slightly tuned the parser (make it support comments at the same line and report errors being more detailed)

* Small tune

* sprite_funcs.cpp: Small tune

* Update lunarender.h

Beautify

* Attempt to fix the MSVC build

* MemEMU: Move value conversions into own calls

Also, resolved the Raocow problem at Science Final Battle. The reason was in the casting of 65545 into int16 as -1. So, the correct algorithm is:
- first off, cast into uint16_t to normally cast the value
- then, cast into int16_t to get the -1

* Don't play sounds oftenly than per 4 frames

* autocode.cpp: Handle the demo counter LunaControl act

* Swap addresses of max stars and num stars

They was been confused between each other, now should be correct

* Update lunadefs.h

Small space added

* MemEMU: Added the "Cheater" address into global table

* Bump luna version into 9

* Use the "LunaScript" term at debug prints

* MemEMU: Organize the fields mapper

* MemEMU: Allow byte hacking of fields

Organize the ability to write single bytes into any of target values with float, double, and integer types

* MemEMU: Allow more deeper byte hacking

* MemEMU: Disable unused code fragment

* MemEMU: Optimize the work of object memory module

Do use normal arrays here, unordered map is overhead here as all addresses of the range were well known

* Ouch

Added missing #include <cstddef> include

* MemEMU: Tiny fix

* MemEMU: More global addresses added

* MemEMU: Sort global fields by address order

* Added variables bank through SAVX file

Repaired LunaDLL's saved variables bank, now it gets saved and loaded at the SAVX file

* Copyright year update for new modules

* Added an ability to disable/enable Luna Autocode build

I made the `-DTHEXTECH_ENABLE_LUNA_AUTOCODE=ON/OFF`

* MemEMU: Fixed the invalid assert at boolean field check

* Make the sprite API partially work

I am not sure, was it actually used anywhere or not, because it was not documented and it was a sort of the experimental thing in the past, or something also.

* Fixed LunaImage build

* Update AudioCodecs

* Tweaked version printing out

Added the revision number

* Set version into 1.3.6-dev

* Android: Get string from CMake side

It's no longer needed to specify the version manually

* Android: Bump modules versions

And small clean-up

* Android: Organize the version code

TODO: Hook the build number too

* Updated submodules

(and attempt to fix the Emscripten build)

* MemEmu: Fixed compatibility after update

Fixed compatibility with the updated set of global variables (GamePaused is now enum, and needs the lambda-function to mimicry as old boolean value)