Canonical vScreens now respect the qScreen offset
This may still be a temporary fix. It provably works at SMBX resolution, and it displays the right behavior on my test cases, but if any other bugs pop up, I may decide to totally replicate the entire vScreen system (vScreen, vScreenX/Y, qScreen, qScreenX/Y, qScreenLoc) with canonical objects and duplicate all of the code that touches it. That would be hard to maintain!