Added an option: Don't grab shells from top fall
This option disables an ability of player to grab a shell while stomping it from the top and holds the RUN button. It's very importand option to prevent soft-locks in ported legacy episodes where player can accidentally destroy a shell hidden in the niche which is required to continue the path.