Fixed the silent music and FX at warp-only levels
When these levels has no start points, and the initial logic fails to find the section where player should appear, and when section is set to 0, it already matches the automatic choice. So, no music starts. However, at warp entrance installation, when I set the -1 to section, it successfully gets found and everything works as expected. Previously this was powered by a "black magic" of unused section that caused section id to mismatch in the condition at final.